refactor to build virtualenvs with a custom step
The long shell scripts were getting out of hand.  This uses virtualenv
and pip in such a way that (almost) all downloads come from the
buildmaster, rather than from pypi (which sometimes causes failures).
This will also nail down the versions used in testing, at the expense of
needing to manually upgrade them.

Also includes a conversion of tabs to spaces, so the diff is longer than
necessary.