rework pip install dependency scheme
pip is not a dependency resolver.

when using pip install, it only looks at the first contraint it sees for each package

for example if you do

pip install twisted -e ./master

pip will ignore the constraint on twisted version that is on master/setup.py

then python will crash when trying to use buildbot module:

pkg_resources.DistributionNotFound: The 'Twisted<16.3.0,>=14.0.1' distribution was not found and is requ
ired by buildbot

So if we are not constraining the dependencies version, we let the one defined in master/setup.py take the precedence.

We also need to make sure slave is last as it has an unconstraint dependency to twisted